requirements.txt
# SmartResume AI — Dependencies
streamlit>=1.41.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
PyPDF2>=3.0.0
python-docx>=1.1.0
numpy>=1.24.0
nltk>=3.8.0
spacy>=3.7.0
